+++ src/app/(tabs)/story/index.tsx
@@ -1,4 +1,3 @@
-import { useLocalSearchParams, useRouter } from 'expo-router';
 import React, { useEffect, useState } from 'react';
 import {
   ActivityIndicator,
@@ -20,10 +19,6 @@ function StoryScreen() {
   const [isLoading, setLoading] = useState(true);
   const scrollRef = React.useRef<any>(null);
   const [story, setStory] = useState<any>();
-
-  const params = useLocalSearchParams();
-  const { author } = params;
-  const router = useRouter();
 
   const onShare = async () => {
     try {
@@ -66,7 +61,6 @@ function StoryScreen() {
           <Image style={styles.image} source={{ uri: story.featured_media }} />
 
           <Text style={styles.title}>{story.title}</Text>
-          <Text>{author}</Text>
 
           <View style={styles.author}>
             <Image style={styles.authorImage} source={{ uri: '' }} />
