+++ src/components/SplashScreen/styles.ts
@@ -1,0 +1,16 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: 'white',
+  },
+  logo: {
+    width: 200,
+    height: 106.81,
+  },
+});
+
+export default styles;
